seg fault avl

+++ hashtable.hpp
@@ -37,7 +37,7 @@ private:
 public:
     HashTable(const int);
     ~HashTable();
-    int hashFunction(Entry*);
+    unsigned long long hashFunction(Entry*);
     void insert(Entry*);
 };
